Refactor Zwift services and improve data handling
Refactored `ZwiftEventsService` to improve modularity by moving event downloading logic to a new `ZwiftStoreEventsService` class. Added methods for validating event URLs, checking race completion, and skipping unfinished events.

Enhanced `GetEntrants` to support an `includeMyself` parameter and return a new `ZwiftEntrantResponseModel` with updated property types. Updated data models to use `long` for IDs and `uint` for non-negative values like `Age`, `Weight`, and `Ftp`.

Improved code readability by refactoring methods like `GetRiderHistory` and `GetWkgValue`. Introduced `FitSyncHub.Common.Extensions` for null filtering. Updated weight calculations to use `WeightInGrams` for consistency.

These changes enhance maintainability, readability, and functionality.

+++ src/FitSyncHub.Zwift/HttpClients/Models/Responses/Events/ZwiftEventSubgroupEntrantResponse.cs
@@ -2,17 +2,17 @@
 
 public record ZwiftEventSubgroupEntrantResponse
 {
-    public required int Id { get; init; }
+    public required long Id { get; init; }
     public required string PublicId { get; init; }
     public required string FirstName { get; init; }
     public required string LastName { get; init; }
     public required bool Male { get; init; }
     public required string EventCategory { get; init; }
-    public required int Age { get; init; }
+    public required uint Age { get; init; }
     public required int BodyType { get; init; }
     public required int Height { get; init; }   // in mm
-    public required int Weight { get; init; }   // in grams
-    public required int Ftp { get; init; }
+    public required uint Weight { get; init; }   // in grams
+    public required uint Ftp { get; init; }
     public required int AchievementLevel { get; init; }
     public required long TotalDistance { get; init; }          // meters
     public required long TotalDistanceClimbed { get; init; }   // meters
